requirements.txt
pandas>=1.3.0
Pillow>=8.0.0
openpyxl>=3.0.0